feat: new drawer menu
