+++ assets/icons/navigation/index.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import MyStudy from './my-study.svg';
+import EditProfiles from './edit-profiles.svg';
+import Share from './share.svg';
+import VaccineRegistry from './vaccine-registry.svg';
+
+export const MyStudyIcon: React.FC = () => <MyStudy height={11.25  * 2} width={14.38 * 2}/>;
+export const EditProfilesIcon: React.FC = () => <EditProfiles height={10  * 2} width={10 * 2}/>;
+export const ShareIcon: React.FC = () => <Share height={12  * 2} width={12 * 2}/>;
+export const VaccineRegistryIcon: React.FC = () => <VaccineRegistry height={12  * 2} width={12 * 2}/>;
