+++ assets/lang/en.json
@@ -565,6 +565,8 @@
   "research-updates": "Research updates",
   "faqs": "Frequently asked questions",
   "privacy-policy": "Privacy policy",
+  "nav-edit-profile": "Edit profile",
+  "nav-share-this-app": "Share this app",
   "delete-my-data": "Delete my data",
   "delete-data-alert-title": "Are you sure?",
   "delete-data-alert-text": "Deleting your account and data cannot be undone.",
