+++ assets/lang/es.json
@@ -562,6 +562,8 @@
   "research-updates": "Noticias de la investigación",
   "faqs": "Preguntas frecuentes",
   "privacy-policy": "Política de privacidad",
+  "nav-edit-profile": "Editar perfil",
+  "nav-share-this-app": "Comparte esta aplicación",
   "delete-my-data": "Eliminar mis datos",
   "delete-data-alert-title": "¿Está seguro?",
   "delete-data-alert-text": "Si elimina su cuenta y datos, no se puede deshacer.",
