+++ assets/lang/sv-SE.json
@@ -490,6 +490,8 @@
   "research-updates": "Aktuella forskningsnyheter",
   "faqs": "Vanliga frågor",
   "privacy-policy": "Integritetsmeddelande",
+  "nav-edit-profile": "Redigera profil",
+  "nav-share-this-app": "Dela denna appen",
   "delete-my-data": "Ta bort mina uppgifter",
   "delete-data-alert-title": "Är du säker?",
   "delete-data-alert-text": "Du kan inte ångra dig om du tar bort ditt konto och dina uppgifter.",
