+++ src/components/Text.tsx
@@ -62,7 +62,7 @@ export const ClickableText = ({ style, children, onPress }: ClickableProps) => (
   </Text>
 );
 
-export const Divider = () => <View style={styles.divider} />;
+export const Divider: React.FC<{styles?: StyleProp<ViewStyle>}> = ({ styles: passed }) => <View style={[styles.divider, passed]} />;
 
 const styles = StyleSheet.create({
   headerText: {
